Prepare for further OctreeNode reinsertions during threaded update phase.

+++ Turso3D/Thread/WorkQueue.cpp
@@ -114,7 +114,7 @@ void WorkQueue::Complete()
 
     for (;;)
     {
-        size_t tasksLeft = numPendingTasks.load();
+        int tasksLeft = numPendingTasks.load();
         if (!tasksLeft)
             break;
